Add tough_ce, add remaining changes for lr scheduler and env wrapper.

+++ alphazero/cpp_impl/game.hpp
@@ -4,13 +4,13 @@
 
 class Game {
   public:
-    std::vector<std::vector<int>> states;
+    std::vector<std::vector<float>> states;
     std::vector<double> rewards;
     std::vector<std::vector<double>> mcts_actions;
 
     Game() {};
     Game(
-      std::vector<std::vector<int>> states,
+      std::vector<std::vector<float>> states,
       std::vector<double> rewards,
       std::vector<std::vector<double>> mcts_actions
     ) : states(states), rewards(rewards), mcts_actions(mcts_actions) {};
